requirements.txt
discord.py
python-dotenv
yfinance
pandas
numpy
matplotlib
scipy
pandas-datareader
seaborn
asyncio
textract==1.6.3
xxhash
ratelimit
pydantic
instructor
openai